requirements.txt
<<<<<<< HEAD
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozendict==2.4.4
idna==3.7
lxml==5.2.2
peewee==3.17.5
pytz==2024.1
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
=======
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
fonttools==4.53.0
frozendict==2.4.4
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
peewee==3.17.5
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.1
uritemplate==4.1.1
>>>>>>> c3d67408cda330c6b1850ebe0fb3eed37120c0c8
urllib3==2.2.1
yahoofinancials==1.20
